Refactor callbacks controller
By splitting it into multiple controllers: one per callback.
