+++ config/dev.exs
@@ -38,7 +38,7 @@ config :ex_money, ExMoney.Repo,
   pool_size: 10
 
 config :ex_money, :login_logger_worker,
-  [name: :login_logger, enabled: true]
+  [name: ExMoney.Saltedge.LoginLogger, enabled: true]
 
 # Put Saltedge credentials into dev.secret.exs
 # config :ex_money,
